requirements.txt
bcrypt==4.3.0
blinker==1.9.0
cffi==1.17.1
click==8.2.2
cryptography==45.0.5
Flask==3.1.1
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
python-dotenv==1.1.1
Werkzeug==3.1.3
